refactor: move single project configuration to a separate object
