+++ ij-core/src/main/kotlin/org/jetbrains/research/ij/headless/server/utils/PsiFileUtils.kt
@@ -6,6 +6,9 @@ import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
+import com.intellij.psi.search.FilenameIndex
+import com.intellij.psi.search.GlobalSearchScope
+import org.jetbrains.kotlin.idea.core.util.toPsiFile
 
 fun updatePsiFileContent(psiFile: PsiFile, text: String) {
     val document = psiFile.viewProvider.document
@@ -19,4 +22,14 @@ fun updatePsiFileContent(psiFile: PsiFile, text: String) {
 fun createPsiFile(project: Project, fileName: String, language: Language, text: String = ""): PsiFile {
     return PsiFileFactory.getInstance(project).createFileFromText(fileName, language, text)
         ?: error("Can not create file $fileName")
+}
+
+fun findPsiFileByName(project: Project, fileName: String): PsiFile {
+    val scope = GlobalSearchScope.allScope(project)
+    
+    val mainVirtualFile = FilenameIndex.getVirtualFilesByName(fileName, scope).firstOrNull() ?: run {
+        error("Can not find the $fileName file in the ${project.name} project")
+    }
+
+    return mainVirtualFile.toPsiFile(project)!!
 }
